backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pymysql
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
